requirements.txt
requests[socks]
rich
googlesearch-python